Combine files with long and short file names
Previously we emitted an inode on each of the filenames of an MFT
entry. This is confusing because both refer to the same file. However
malicious users may rename the short file name to something different
and most tools only show the long file name. The risk is that a user
may discover a reference to a file but not be able to find the file -
thinking it was removed. Actually the file may exist but will have a
LFN different from the short file name it sees.

This change hides the short file name if it contains a ~ which is
usually the case for normal files. If a reference is found to a file
with a ~ in it the user may expect this to be a short file name
anyway. So it removed confusion. But in the common case the short file
names are not relevant.