Added support for multiple vcns. (#25)
Files may contain multiple VCN ranges in which case the MFT entry
contains multiple streams with the same type and id but representing
multiple ranges over the same file. This change adds support for
these.

The change also introduces an new Ranges() method which reports all
the ranges of a stream including if the range is sparse or not. This
allows callers to skip sparse ranges efficiently.

* Added testing against TSK's icat.

Opens each stream and calculates hashes then compares to the icat hash.