Added error reporting when reconstructing the path (#57)
Sometimes when reconstructing MFT paths we hit errors (especially for
unallocated entried). The previous code would just stop the path
reconstruction at the point of error and since paths are built in
reverse it would appear like the last valid directory is actually
rooted at the top of the filesystem.

This change creates special "psuedo" directories where these paths can
terminate making it clear that they are errors. We also add extra
components describing the error of why we were unable to continue
processing.

Also added sequence number check when reconstructing paths - will now reject a parent MFT entry if the sequence numbers do not match the one specified in the $FILENAME attribute. This check might be too stringent on unallocated files but it seems to give no FP on allocated files in practice.