Fix reading beyond the end of a file (#101)
This PR addresses an issue in the PagedReader that can lead to an
infinite loop at the end of a file, eventually causing an out-of-memory
condition. The problem occurs because, in some cases, the code invoking
the PagedReader cannot reliably detect when the end of the file has been
reached. This behavior can be reproduced using the Velociraptor query
below.

```
SELECT * FROM parse_jsonl(filename='''C:\test\test.jsonl''', accessor='raw_file')
```

Rendered pprof CPU profile:

![profile](https://github.com/user-attachments/assets/59c43b20-35c7-4006-ab43-e2c666df9dd3)

---------

Co-authored-by: Mike Cohen <mike@velocidex.com>